refactor: construct notifications in operations module
Move the action that constructs Notifications to the N.Operations
module, and rename it to makeNotification. N.Main has also been updated
to use the operations module now that a couple symbols were removed from
N.Core.

Notable addition: a lens into the ID of a notification to be replaced
received from DBus was added to N.Lenses.